changes to plottong directory 	modified:   filelist/list_2018_MC_A01.txt 	modified:   kraken_run.py 	modified:   monitor.py 	renamed:    plotting/make_plots_MC.py -> plotting/make_plots.py 	deleted:    plotting/make_plots_data.py 	modified:   plotting/plot_MC.ipynb 	new file:   plotting/plot_all.py 	deleted:    plotting/plot_all_MC.py 	deleted:    plotting/plot_all_data.py 	modified:   plotting/plot_data.ipynb 	modified:   workflows/SUEP_coffea.py
